// types/themeType.ts
export type themeType = {
    PRIMARY: string;
    BACKGROUND: string;
    NAVIGATION: string;
    TEXT: string;
    DISABLED: string;
    CARD: string;
    ICONS: string;
    DELETE: string;
    PLACEHOLDER: string;
    ANGRY: { BACKGROUND: string; TEXT: string };
    FANTASTIC: { BACKGROUND: string; TEXT: string };
    GREAT: { BACKGROUND: string; TEXT: string };
    HAPPY: { BACKGROUND: string; TEXT: string };
    NEUTRAL: { BACKGROUND: string; TEXT: string };
    SAD: { BACKGROUND: string; TEXT: string };
    STRESSED: { BACKGROUND: string; TEXT: string };
  };
  